// src/app/components/admin-ho/vendor/new_vendor/types.ts
export interface StoreSelect {
  value: number
  label: string
}

export interface ServiceArea {
  value: any
  label: string
}

export interface ServiceType {
  value: any
  label: string
}

export interface Bank {
  value: any
  label: string
}

export interface CheckStates {
  compro: boolean
  suratPermohonan: boolean
  pks: boolean
  suip: boolean
  ptkp: boolean
}

export interface ImagePreview {
  blob: string
  fileName: string
}